refactor(contracts): fix circular dependency between container and pipelines/contracts
Move PipelineContainerABC and PipelineFactoryABC from
application/pipelines/contracts.py to application/contracts.py.
This breaks the circular import between container.py and pipelines/contracts.py.

Changes:
- Create new application/contracts.py with core pipeline abstractions
- Update pipelines/contracts.py to re-export for backwards compatibility
- Update all imports across the codebase to use new location
- Export contracts from application/__init__.py
