+++ src/bioetl/application/pipelines/__init__.py
@@ -4,11 +4,8 @@ Pipelines package.
 Provides base pipeline implementation, contracts, and factory registry.
 """
 
+from bioetl.application.contracts import PipelineContainerABC, PipelineFactoryABC
 from bioetl.application.pipelines.base import PipelineBase
-from bioetl.application.pipelines.contracts import (
-    PipelineContainerABC,
-    PipelineFactoryABC,
-)
 from bioetl.application.pipelines.registry import (
     get_factory,
     get_pipeline_factory,
